+++ src/bract/ring/config.clj
@@ -10,15 +10,28 @@
 (ns bract.ring.config
   (:require
     [keypin.core :as keypin]
-    [keypin.util :as kputil]))
+    [keypin.util :as kputil]
+    [bract.core.echo :as echo]))
 
 
-(keypin/defkey
-  ;; context keys
-  ctx-ring-handler [:bract.ring/ring-handler fn?     "Application Ring handler"]
-  ;; config keys
-  cfg-wrappers     ["bract.ring.wrappers"    vector? "Fully qualified wrapper fn names"
-                    {:parser kputil/any->edn}])
+(keypin/defkey  ; context keys
+  ctx-ring-handler [:bract.ring/ring-handler fn? "Application Ring handler"])
+
+
+(keypin/defkey  ; config keys
+  cfg-wrappers-context ["bract.ring.wrappers.context" vector? "Fully qualified wrapper (by context) fn names"
+                        {:parser kputil/any->edn}]
+  cfg-wrappers-config  ["bract.ring.wrappers.config"  vector? "Fully qualified wrapper (by config) fn names"
+                        {:parser kputil/any->edn}])
+
+
+(defn apply-wrapper-by-name
+  [wrapper-type config-key handler context wrapper-name]
+  (echo/echo (format "Looking up %s '%s'" wrapper-type wrapper-name))
+  (let [f (kputil/str->var->deref config-key wrapper-name)]
+    (echo/echo (format "Executing  %s '%s'" wrapper-type wrapper-name))
+    (echo/with-inducer-name wrapper-name
+      (f handler context))))
 
 
 (defn update-handler