+++ src/bract/ring/inducer.clj
@@ -15,12 +15,30 @@
     [bract.ring.config :as config]))
 
 
-(defn apply-wrappers
-  "Given a context, look up the ring-handler inducers and apply them successively returning an updated context."
+(defn apply-wrappers-with-context
+  "Given a context with Ring handler, look up the ring-handler wrappers `(fn [handler context]) -> handler` and apply
+  them successively to the handler, finally updating the context with the handler before returning it."
+  ([context]
+    (apply-wrappers-with-context
+      "Ring wrapper (fn [handler context])" config/cfg-wrappers-context identity context))
+  ([wrapper-type config-def f context]
+    (bc-echo/echo "Applying Ring-middleware")
+    (let [handler  (config/ctx-ring-handler context)]
+      (->> (bc-config/ctx-config context)
+        config-def
+        (bc-util/induce handler (fn [updated-handler wrapper-name]
+                                  (config/apply-wrapper-by-name
+                                    wrapper-type
+                                    (key config-def)
+                                    updated-handler
+                                    (f context)
+                                    wrapper-name)))
+        (assoc context (key config/ctx-ring-handler))))))
+
+
+(defn apply-wrappers-with-config
+  "Given a context with Ring handler, look up the ring-handler wrappers `(fn [handler config]) -> handler` and apply
+  them successively to the handler, finally updating the context with the handler before returning it."
   [context]
-  (bc-echo/echo "Applying Ring-handler inducers")
-  (config/ctx-ring-handler context)  ; assert the handler is present
-  (->> (bc-config/ctx-config context)
-    config/cfg-wrappers
-    (bc-util/induce context (partial bc-config/apply-inducer-by-name
-                              "Ring-handler inducer" (key config/cfg-wrappers)))))
+  (apply-wrappers-with-context
+    "Ring wrapper (fn [handler config])" config/cfg-wrappers-config bc-config/ctx-config context))