+++ test/bract/ring/inducer_test.clj
@@ -10,6 +10,7 @@
 (ns bract.ring.inducer-test
   (:require
     [clojure.test :refer :all]
+    [bract.core.config :as bc-config]
     [bract.ring.inducer :as inducer]))
 
 
@@ -17,18 +18,33 @@
 
 
 (defn sample-wrapper
-  [context]
-  (vswap! sample-value (fn [^long n] (+ n 10)))
-  (assoc context :foo :bar))
+  [handler context]
+  (vswap! sample-value (fn [^long n] (+ n 10))))
 
 
-(deftest test-apply-wrappers
+(deftest test-apply-wrappers-context
   (testing "happy cases"
     (let [good-context {:bract.ring/ring-handler identity
-                        :bract.core/config {"bract.ring.wrappers" ['bract.ring.inducer-test/sample-wrapper]}}]
+                        :bract.core/config {"bract.ring.wrappers.context" ['bract.ring.inducer-test/sample-wrapper]}}]
+      (vreset! sample-value 10)
       (is (= 10 @sample-value))
-      (is (= (assoc good-context :foo :bar)
-            (inducer/apply-wrappers good-context)))
+      (let [context (inducer/apply-wrappers-with-context good-context)]
+        (is (contains? context :bract.ring/ring-handler))
+        (is (contains? context :bract.core/config)))
       (is (= 20 @sample-value))))
   (testing "missing/bad keys"
-    (is (thrown? IllegalArgumentException (inducer/apply-wrappers {})))))
+    (is (thrown? IllegalArgumentException (inducer/apply-wrappers-with-context {})))))
+
+
+(deftest test-apply-wrappers-config
+  (testing "happy cases"
+    (let [good-context {:bract.ring/ring-handler identity
+                        :bract.core/config {"bract.ring.wrappers.config" ['bract.ring.inducer-test/sample-wrapper]}}]
+      (vreset! sample-value 10)
+      (is (= 10 @sample-value))
+      (let [context (inducer/apply-wrappers-with-config good-context)]
+        (is (contains? context :bract.ring/ring-handler))
+        (is (contains? context :bract.core/config)))
+      (is (= 20 @sample-value))))
+  (testing "missing/bad keys"
+    (is (thrown? IllegalArgumentException (inducer/apply-wrappers-with-config {})))))